Add files via upload
SQL-Analyze-International-Debt-Statistics
Analyzing an international debt data collected by The World Bank and finding the answers to questions like:

What is the total amount of debt that is owed by the countries listed in the dataset?
Which country owns the maximum amount of debt and what does that amount look like?
What is the average amount of debt owed by countries across different debt indicators?
The data used in this project is provided by The World Bank. It contains both national and regional debt statistics for several countries across the globe as recorded from 1970 to 2015. https://datatopics.worldbank.org/debt/ids/country/CHN